pyproject.toml
[project]
name = "dragonwizard"
version = "0.1.0"
description = "LLM-powered Discord bot for D&D 5th edition rules assistance"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "mcp>=1.0.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.3.0",
    "pymupdf>=1.23.0",
    "tiktoken>=0.5.2",
    "nltk>=3.8.1",
    "aiofiles>=23.2.1",
    "numpy>=1.24.0",
    "litellm>=1.0.0",
    "discord-py>=2.0",
]

[project.scripts]
dragonwizard = "dragonwizard.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]
